username removed added email

+++ ekart-backend/Order-Service/src/main/java/com/ekart/order/entity/PromoCodes.java
@@ -18,7 +18,7 @@ public class PromoCodes {
 	@Id
 	private int promoCodeId;
 	private String promoCode;
-	private int discountPrice;
+	private double discountPrice;
 	private String status;
 
 }
